Drop icons
For the system check using XH_message() (or a wrapper) is customary
anyway.  For the warning about desynchronized page data, we use a
Unicode warning symbol for now – this check can likely be dropped
anyway, as since CMSimple_XH 1.5 the issue has likely been solved,
and as of CMSimple_XH 1.6 it is almost certainly.  Only if users
edit content.htm manually and introduce a bug there, the check may
still make sense.  We decide upon that later.