Test MainAdminController through more real Model
Testing controllers against a model mock is barely useful; at the very
least we should test against a model where the dependencies are mocked.
It would be even better to only partially mock the dependencies, but
the classes do not allow this without considerable work, so we stick
with this for now.  Since the actual model test is now a duplicate, we
drop it altogether.